+++ lib/src/spec/object_field_selector.g.dart
@@ -7,9 +7,10 @@ part of 'object_field_selector.dart';
 // **************************************************************************
 
 ObjectFieldSelector _$ObjectFieldSelectorFromJson(Map<String, dynamic> json) =>
-    ObjectFieldSelector()
-      ..apiVersion = json['apiVersion'] as String
-      ..fieldPath = json['fieldPath'] as String;
+    ObjectFieldSelector(
+      apiVersion: json['apiVersion'] as String?,
+      fieldPath: json['fieldPath'] as String?,
+    );
 
 Map<String, dynamic> _$ObjectFieldSelectorToJson(
         ObjectFieldSelector instance) =>