+++ lib/src/spec/volume/projection/downward_api_projection.g.dart
@@ -8,14 +8,12 @@ part of 'downward_api_projection.dart';
 
 DownwardAPIProjection _$DownwardAPIProjectionFromJson(
         Map<String, dynamic> json) =>
-    DownwardAPIProjection()
-      ..items = (json['items'] as List<dynamic>?)
-          ?.map(
-              (e) => DownwardAPIVolumeFile.fromJson(e as Map<String, dynamic>))
-          .toList();
+    DownwardAPIProjection(
+      items: _itemsFromJson(json['items'] as List?),
+    );
 
 Map<String, dynamic> _$DownwardAPIProjectionToJson(
         DownwardAPIProjection instance) =>
     <String, dynamic>{
-      if (instance.items case final value?) 'items': value,
+      if (_itemsToJson(instance.items) case final value?) 'items': value,
     };