+++ lib/src/spec/volume/projection/secret_projection.g.dart
@@ -7,16 +7,15 @@ part of 'secret_projection.dart';
 // **************************************************************************
 
 SecretProjection _$SecretProjectionFromJson(Map<String, dynamic> json) =>
-    SecretProjection()
-      ..items = (json['items'] as List<dynamic>?)
-          ?.map((e) => KeyToPath.fromJson(e as Map<String, dynamic>))
-          .toList()
-      ..name = json['name'] as String?
-      ..optional = json['optional'] as bool?;
+    SecretProjection(
+      items: _itemsFromJson(json['items'] as List?),
+      name: json['name'] as String?,
+      optional: json['optional'] as bool?,
+    );
 
 Map<String, dynamic> _$SecretProjectionToJson(SecretProjection instance) =>
     <String, dynamic>{
-      if (instance.items case final value?) 'items': value,
+      if (_itemsToJson(instance.items) case final value?) 'items': value,
       if (instance.name case final value?) 'name': value,
       if (instance.optional case final value?) 'optional': value,
     };